% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{div_b}
\alias{div_b}
\title{div_b Function}
\usage{
div_b(bse, x, y)
}
\arguments{
\item{bse}{integer from 2 to 9. base of the numbers.}

\item{x}{number in base bse}

\item{y}{number in base bse}
}
\value{
division in base bse
}
\description{
this functions does integer division in a given base. if division is not integer, the result may be inexact.
}
\examples{
div_b(5,30,10)
div_b(5,20,2)
}
\keyword{base}
