% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{div_int}
\alias{div_int}
\title{div_int Function}
\usage{
div_int(bse, x)
}
\arguments{
\item{bse}{integer}

\item{x}{integer}
}
\value{
integer
}
\description{
this function performs integer division in a given base. Used by from10_to_x function. not intended for user use.
}
\keyword{base}
