% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{factorize_bse}
\alias{factorize_bse}
\title{factorize_bse Function}
\usage{
factorize_bse(bse, x)
}
\arguments{
\item{bse}{integer from 2 to 9. base of the numbers.}

\item{x}{number in base bse to be decomposed}
}
\value{
vector with primes
}
\description{
this functions generates the prime decomposition of a number in a given base
}
\examples{
factorize_bse(5,30)
factorize_bse(5,20)
}
\keyword{base}
