% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{from10_to_x}
\alias{from10_to_x}
\title{from10_to_x Function}
\usage{
from10_to_x(bse, x)
}
\arguments{
\item{bse}{integer from 2 to 9. New base of the number.}

\item{x}{number in base 10}
}
\value{
integer in base bse
}
\description{
this functions converts integers in base 10 to a given base.
}
\examples{
from10_to_x(2,7)
from10_to_x(5,26)
}
\keyword{base}
