% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{from_x_to_y}
\alias{from_x_to_y}
\title{from_x_to_y Function}
\usage{
from_x_to_y(bse1, bse2, x)
}
\arguments{
\item{bse1}{integer from 2 to 10. Original base of the number}

\item{bse2}{integer from 2 to 10. New base to convert the number}

\item{x}{number in base bse}
}
\value{
integer in base 10
}
\description{
this functions converts integers from a given base to another
}
\examples{
###convert 44 in base 5 to base 3
from_x_to_y(5,3,44)
}
\keyword{base}
