% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{fromx_to_10}
\alias{fromx_to_10}
\title{fromx_to_10 Function}
\usage{
fromx_to_10(bse, x)
}
\arguments{
\item{bse}{integer from 2 to 9. Original base of the number}

\item{x}{number in base bse}
}
\value{
integer in base 10
}
\description{
this functions converts integers from a given base to base 10
}
\examples{
fromx_to_10(2,1101)
fromx_to_10(5,43)
}
\keyword{base}
