% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{prod_b}
\alias{prod_b}
\title{prod_b Function}
\usage{
prod_b(bse, x, y)
}
\arguments{
\item{bse}{integer from 2 to 9. base of the numbers.}

\item{x}{number in base bse}

\item{y}{number in base bse}
}
\value{
product in base bse
}
\description{
this functions does product in a given base.
}
\examples{
prod_b(2,10,101)
prod_b(5,24,23)
}
\keyword{base}
