% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{random_bin}
\alias{random_bin}
\title{random_bin Function}
\usage{
random_bin(bse, dgts, op)
}
\arguments{
\item{bse}{integer from 2 to 9. base of the numbers.}

\item{dgts}{number of digits of the numbers to be generated}

\item{op}{character either "sum" or "prod"}
}
\value{
division in base bse
}
\description{
this functions generates random exercises for prod and sum
}
\examples{
div_b(5,30,10)
div_b(5,20,2)
}
\keyword{base}
