% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refous.R
\name{sum_b}
\alias{sum_b}
\title{sum_b Function}
\usage{
sum_b(bse, x, y)
}
\arguments{
\item{bse}{integer from 2 to 9. base of the numbers.}

\item{x}{number in base bse}

\item{y}{number in base bse}
}
\value{
sum in base bse
}
\description{
this functions does sum in a given base.
}
\examples{
sum_b(2,10,101)
sum_b(5,24,23)
}
\keyword{base}
